% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeBase.R
\docType{data}
\name{ce_v12_08_NetGPA}
\alias{ce_v12_08_NetGPA}
\title{Coexpression network}
\format{a integer matrix which contains top 10 percent most significant edges}
\usage{
data(ce_v12_08_NetGPA)
}
\description{
Coexpression network
}
\keyword{datasets}
