% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeBase.R
\docType{data}
\name{DEPICT_2015_01_NetGPA}
\alias{DEPICT_2015_01_NetGPA}
\title{A integrative network}
\format{a integer matrix which contains top 10 percent most significant edges}
\usage{
data(DEPICT_2015_01_NetGPA)
}
\description{
A integrative network
}
\keyword{datasets}
