% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeBase.R
\docType{data}
\name{Example_NetGPA}
\alias{Example_NetGPA}
\title{Example data set used in NetGPA}
\format{check vignettes for details}
\usage{
data(Example_NetGPA)
}
\description{
Example data set used in NetGPA
}
\keyword{datasets}
