% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneWeight.R
\name{geneWeight}
\alias{geneWeight}
\title{Calculate gene weight}
\usage{
geneWeight(queryIndex, Nref, topIndex, seedList, Pfcutoff)
}
\arguments{
\item{queryIndex}{The index of query gene.}

\item{Nref}{Total number of genes in the database.}

\item{topIndex}{Top indexes that are associated with queryIndex gene}

\item{seedList}{A list of intger vectors, each of which represents indexes of seed genes.}

\item{Pfcutoff}{Percentage of top genes considered, should be less than 0.1.}
}
\value{
A vector with 2 variables, Nseeds and Cg.
}
\description{
This function is the core function to calculate weight score for a gene.
}
\examples{
# load package and data
library("NetGPA")
data("Example_NetGPA")

# load data base
data(text_2006_12_NetGPA)

# Compare FGF pathway and a random gene set
exampleSeedFGFR <- Example_NetGPA$Cancer_GeneSet[["SIGNALING_BY_FGFR"]]

pGenes <- intersect(exampleSeedFGFR, colnames(text_2006_12_NetGPA))
rGenes <- sample(colnames(text_2006_12_NetGPA), length(pGenes)) 

res    <- NetGPA(as.list(pGenes), pGenes, text_2006_12_NetGPA, progressBar = FALSE)
pTable <- res$queryTable
pTable$group <- "Pathway"

res    <- NetGPA(as.list(rGenes), rGenes, text_2006_12_NetGPA, progressBar = FALSE)
rTable <- res$queryTable
rTable$group <- "Random"

mT <- rbind(pTable, rTable)

boxplot(-log10(queryP)~group, data = mT, ylab = "-log10(p-value)")

}
\references{
Jiantao Shi: NetGPA, a package for network-based gene prioritization analysis.
}
\author{
Jiantao Shi
}
