% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexSeedList.R
\name{indexSeedList}
\alias{indexSeedList}
\title{Calculate gene weight}
\usage{
indexSeedList(seedListName, dbMatrix)
}
\arguments{
\item{seedListName}{A list of vectors which contain gene names in each region.}

\item{dbMatrix}{Database matrix}
}
\value{
A list of vectors which contain gene indexes in each region.
}
\description{
This function is the core function to calculate weight score for a gene.
}
\examples{
# load package and data
library("NetGPA")
data("Example_NetGPA")

# load data base
data(text_2006_12_NetGPA)

# Compare FGF pathway and a random gene set
exampleSeedFGFR <- Example_NetGPA$Cancer_GeneSet[["SIGNALING_BY_FGFR"]]

pGenes <- intersect(exampleSeedFGFR, colnames(text_2006_12_NetGPA))
rGenes <- sample(colnames(text_2006_12_NetGPA), length(pGenes)) 

res    <- NetGPA(as.list(pGenes), pGenes, text_2006_12_NetGPA, progressBar = FALSE)
pTable <- res$queryTable
pTable$group <- "Pathway"

res    <- NetGPA(as.list(rGenes), rGenes, text_2006_12_NetGPA, progressBar = FALSE)
rTable <- res$queryTable
rTable$group <- "Random"

mT <- rbind(pTable, rTable)

boxplot(-log10(queryP)~group, data = mT, ylab = "-log10(p-value)")

}
\references{
Jiantao Shi, Franziska Michor, Winston Hide: NetGPA, a package for network-based gene prioritization analysis.
}
\author{
Jiantao Shi
}
