% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGPA.R
\name{NetGPA}
\alias{NetGPA}
\title{NetGPA}
\usage{
NetGPA(seedListName, queryName, dbMatrix, Pfcutoff = 0.1,
  progressBar = TRUE)
}
\arguments{
\item{seedListName}{A list which contains the seed gene set.}

\item{queryName}{A vector of gene names.}

\item{dbMatrix}{Network database in a matrix format.}

\item{Pfcutoff}{Percentage of top genes considered, should be less than 0.1.}

\item{progressBar}{Whether to show process status, logical.}
}
\value{
A list of two tables
}
\description{
The workflow to score a group of query genes
}
\examples{
# load package and data
library("NetGPA")
data("Example_NetGPA")

# load data base
data(text_2006_12_NetGPA)

# Compare FGF pathway and a random gene set
exampleSeedFGFR <- Example_NetGPA$Cancer_GeneSet[["SIGNALING_BY_FGFR"]]

pGenes <- intersect(exampleSeedFGFR, colnames(text_2006_12_NetGPA))
rGenes <- sample(colnames(text_2006_12_NetGPA), length(pGenes)) 

res    <- NetGPA(as.list(pGenes), pGenes, text_2006_12_NetGPA, progressBar = FALSE)
pTable <- res$queryTable
pTable$group <- "Pathway"

res    <- NetGPA(as.list(rGenes), rGenes, text_2006_12_NetGPA, progressBar = FALSE)
rTable <- res$queryTable
rTable$group <- "Random"

mT <- rbind(pTable, rTable)

boxplot(-log10(queryP)~group, data = mT, ylab = "-log10(p-value)")

}
\references{
Jiantao Shi NetGPA, a package for network-based gene prioritization analysis.
}
\author{
Jiantao Shi
}
