% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeBase.R
\docType{data}
\name{text_2006_12_NetGPA}
\alias{text_2006_12_NetGPA}
\title{Text-mining network}
\format{a integer matrix which contains top 10 percent most significant edges}
\usage{
data(text_2006_12_NetGPA)
}
\description{
Text-mining network
}
\keyword{datasets}
